拆分 is4map6addr 和 getv4addr 函数，用于和 smth_sshbbsd 共用
