+++ src/bbsd/CMakeLists.txt
@@ -1,2 +1,2 @@
-add_executable(bbsd bbsd.c)
+add_executable(bbsd bbsd.c ip.c)
 target_link_libraries(bbsd ythtbbs ytht)
